Websocket/requirements.txt
websocket-client==1.8.0
websockets==13.1
pymongo==4.10.1
apscheduler==3.10.4